# application/api-spec/src/main/resources/openapi/isk.yaml
openapi: 3.0.0
info:
  title: Informačný systém knižnice (ISK)
  version: 1.0.0
  description: API pre správu knižničného systému – katalóg kníh, výpožičky, rezervácie a členstvá.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /books:
    get:
      summary: Získanie zoznamu všetkých kníh
      operationId: getAllBooks
      tags:
        - Book
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
        - name: author
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/BookGenre'
      responses:
        '200':
          description: Zoznam kníh
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Pridanie novej knihy do katalógu
      operationId: createBook
      tags:
        - Book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Kniha úspešne pridaná
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{isbn}:
    get:
      summary: Získanie knihy podľa ISBN
      operationId: getBookByIsbn
      tags:
        - Book
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail knihy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Odstránenie knihy z katalógu
      operationId: deleteBook
      tags:
        - Book
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Kniha úspešne odstránená
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict – kniha má aktívne výpožičky alebo rezervácie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{isbn}/copies:
    post:
      summary: Pridanie nových kópií knihy
      operationId: addBookCopies
      tags:
        - Book
      parameters:
        - name: isbn
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCopiesRequest'
      responses:
        '200':
          description: Kópie úspešne pridané
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── MEMBERS ──────────────────────────────────────────────────────────────
...
  /loans:
    get:
      summary: Získanie zoznamu výpožičiek
      operationId: getAllLoans
      tags:
        - Loan
      parameters:
        - name: memberId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Zoznam výpožičiek
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Vytvorenie novej výpožičky
      operationId: createLoan
      tags:
        - Loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        '201':
          description: Výpožička úspešne vytvorená
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unathorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /loans/overdue:
    get:
      summary: Získanie zoznamu po termíne výpožičiek
      operationId: getOverdueLoans
      tags:
        - Loan
      responses:
        '200':
          description: Zoznam po termíne výpožičiek
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /loans/{id}/return:
    post:
      summary: Vrátenie vypožičanej knihy
      operationId: returnLoan
      tags:
        - Loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Kniha úspešne vrátená
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unathorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /loans/{id}/renew:
    post:
      summary: Predĺženie výpožičky
      operationId: renewLoan
      tags:
        - Loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Výpožička úspešne predĺžená
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unathorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── RESERVATIONS ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
    Money:
      type: object
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: EUR

    # ─── BOOK ───────────────────────────────────────────────────────────────
    Book:
      type: object
      properties:
        isbn:
          type: string
          description: ISBN-10 alebo ISBN-13 identifikátor knihy.
          readOnly: true
        title:
          type: string
        author:
          type: string
        genre:
          $ref: '#/components/schemas/BookGenre'
        publisher:
          type: string
        publicationYear:
          type: integer
        totalCopies:
          type: integer
        availableCopies:
          type: integer
          readOnly: true
        reservedCopies:
          type: integer
          readOnly: true
          description: Počet kópií, ktoré sú aktuálne držané rezerváciami v stave READY_FOR_PICKUP.
    AddCopiesRequest:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          minimum: 1
    CreateBookRequest:
      type: object
      required:
        - isbn
        - title
        - author
        - genre
        - publisher
        - publicationYear
        - totalCopies
      properties:
        isbn:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        title:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        author:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        genre:
          $ref: '#/components/schemas/BookGenre'
        publisher:
          type: string
          minLength: 1
          pattern: "^.*\\S.*$"
        publicationYear:
          type: integer
          minimum: 1
        totalCopies:
          type: integer
          minimum: 1
    BookGenre:
      type: string
      enum:
        - FICTION
        - SCIENCE_FICTION
        - FANTASY
        - MYSTERY
        - THRILLER
        - HORROR
        - ROMANCE
        - HISTORICAL_FICTION
        - ADVENTURE
        - NOVEL
        - BIOGRAPHY
        - AUTOBIOGRAPHY
        - HISTORY
        - SCIENCE
        - TECHNOLOGY
        - PHILOSOPHY
        - PSYCHOLOGY
        - SELF_HELP
        - BUSINESS
        - POLITICS
        - TRUE_CRIME
        - POETRY
        - DRAMA
        - CHILDREN
        - YOUNG_ADULT
        - REFERENCE
        - TRAVEL
        - COOKBOOK
        - ART
        - RELIGION
        - LANGUAGE
        - OTHER

    # ─── MEMBER ─────────────────────────────────────────────────────────────
    Member:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        memberRole:
          $ref: '#/components/schemas/MemberRole'
        membership:
          $ref: '#/components/schemas/Membership'
        fines:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Fine'
...
    Membership:
      type: object
      properties:
        expiryDate:
          type: string
          format: date
          readOnly: true
        status:
          $ref: '#/components/schemas/MembershipStatus'
    Fine:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        amount:
          $ref: '#/components/schemas/Money'
        reason:
          type: string
        status:
          $ref: '#/components/schemas/FineStatus'
    MemberRole:
      type: string
      enum:
        - MEMBER
        - LIBRARIAN
        - ADMIN
    MembershipStatus:
      type: string
      enum:
        - ACTIVE
        - EXPIRED
        - SUSPENDED
    FineStatus:
      type: string
      enum:
        - PENDING
        - PAID
        - WAIVED

    # ─── LOAN ───────────────────────────────────────────────────────────────
    Loan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        loanedTo:
          $ref: '#/components/schemas/Member'
        book:
          $ref: '#/components/schemas/Book'
        createdBy:
          $ref: '#/components/schemas/Member'
        loanDate:
          type: string
          format: date
          readOnly: true
        dueDate:
          type: string
          format: date
          readOnly: true
        returnDate:
          type: string
          format: date
          readOnly: true
          nullable: true
        renewalCount:
          type: integer
          readOnly: true
        status:
          $ref: '#/components/schemas/LoanStatus'
    CreateLoanRequest:
      type: object
      required:
        - memberId
        - isbn
      properties:
        memberId:
          type: integer
          format: int64
          description: ID člena, ktorému sa vystavuje výpožička.
        isbn:
          type: string
          description: ISBN vypožičiavanej knihy.
    LoanStatus:
      type: string
      enum:
        - ACTIVE
        - RETURNED
        - OVERDUE

    # ─── RESERVATION ────────────────────────────────────────────────────────
...